I have updated communitiesList.tsx to have the right and left arrows  be on the same row as "Your communities" and above "You have been added to all communities"
